Uolo-Net/UNet/requirements.txt
absl-py              
astor                
cycler               
decorator            
dicom                
dicompyler-core      
Django                    
django-cors-headers      
docopt                 
gast                  
grpcio              
h5py               
imageio              
Keras               
Keras-Applications      
Keras-Preprocessing  
kiwisolver          
Markdown            
matplotlib           
mock                   
mysqlclient          
networkx             
numpy               
pandas                
patsy                   
pbr                  
Pillow              
pip                     
protobuf             
pydicom              
pykwalify            
pynrrd                
pyparsing            
pyradiomics         
python-dateutil      
pytz                 
PyYAML              
scikit-image         
scikit-learn            
scipy                
setuptools           
SimpleITK               
six                
sklearn             
sqlparse             
tensorboard        
tensorflow-gpu
termcolor            
webcolors             
Werkzeug               
wheel               
xlrd               
